Añadida funcionalidad no poder eliminarse así mismo

+++ Código/Backend/controllers/UserC.js
@@ -1,3 +1,4 @@
+import { BadRequestError } from "../lib/Errors.js"
 import UserService from "../services/UserService.js"
 
 const getUsers = async (req, res) => {
@@ -15,6 +16,9 @@ const postUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
 	const { restauranteId, userId } = req.params
+	if(userId === req.user.id) {
+		throw new BadRequestError("No puedes eliminarte a ti mismo")
+	}
 	await UserService.deleteUser(restauranteId, userId)
 	res.json({ message: "Usuario eliminado"})
 }
